iBATIS is a Mapping Framework for Java POJO
iBATIS is a persistence framework which automates the mapping between
SQL databases and objects in Java, .NET, and Ruby on Rails. In Java, the
objects are POJOs (Plain Old Java Objects). The mappings are decoupled
from the application logic by packaging the SQL statements in XML
configuration files. The result is a significant reduction in the amount
of code that a developer needs to access a relational database using
lower level APIs like JDBC and ODBC